# openapi.yaml
openapi: 3.0.0
info:
  title: Sharded Key-Value Store API
  version: 1.0.0
paths:
  /register_table:
    post:
      summary: Register a new table across all shards
      operationId: coordinator.register_table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table_name:
                  type: string
      responses:
        "201":
          description: Table registered
...